automated_data_pipelines/airflow_weather/requirements.txt
apache-airflow-providers-sqlite
apache-airflow-providers-standard
apache-airflow-providers-fab
apache-airflow-providers-http
apache-airflow-providers-common-sql
requests
pandas
